requirements.txt
-e /Users/sonson/Documents/code/moderntensor
Pillow # Thư viện xử lý ảnh
diffusers
transformers
torch
accelerate
sentencepiece